svg-converter-app/python-backend/requirements.txt
alembic==1.13.1
altgraph==0.17.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
backcall==0.2.0
black==24.8.0
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
comm==0.2.0
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cssselect2==0.7.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
et-xmlfile==1.1.0
executing==2.0.1
flake8==7.1.1
Flask==3.0.3
Flask-Cors==4.0.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
img2pdf==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.27.1
ipython==8.18.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lxml==5.1.0
macholib==1.16.3
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.7.0
minio==7.2.7
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.4
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==24.0
pandas==2.1.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pikepdf==8.11.2
pillow==10.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pyflakes==3.2.0
Pygments==2.17.2
pyinstaller-hooks-contrib==2024.5
pyMEA @ git+https://github.com/kkito0726/MEA_modules.git@c87ca9274235d42eb2c45a5cda83cdd4ff67cc91
PyMySQL==1.1.1
pyodbc==5.1.0
pyparsing==3.1.2
pypdf==3.17.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.1
reportlab==4.2.5
ruff==0.6.5
scikit-learn==1.3.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
svg2csv @ git+https://github.com/kkito0726/svg_converter.git@58cd497c72685062310e7bc0a350cda98f4c98ac
svglib==1.5.1
swan==0.7.3
threadpoolctl==3.2.0
tinycss2==1.4.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.0
tzdata==2023.3
urllib3==2.2.1
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
zipp==3.20.2
